Exception Handling

+++ src/main/java/com/marceloluiz/crudclient/services/ClientService.java
@@ -3,10 +3,15 @@ package com.marceloluiz.crudclient.services;
 import com.marceloluiz.crudclient.dto.ClientDTO;
 import com.marceloluiz.crudclient.entities.Client;
 import com.marceloluiz.crudclient.repositories.ClientRepository;
+import com.marceloluiz.crudclient.services.exceptions.DatabaseException;
+import com.marceloluiz.crudclient.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -21,7 +26,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Resource Not Found"));
         return new ClientDTO(client);
     }
 
@@ -36,15 +42,25 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto){
-        Client client = repository.getReferenceById(id);
-        copyDtoToEntity(client, dto);
+        try{
+            Client client = repository.getReferenceById(id);
+            copyDtoToEntity(client, dto);
 
-        return new ClientDTO(client);
+            return new ClientDTO(client);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)) throw new ResourceNotFoundException("Resource Not Found");
+
+        try{
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Referential Integrity Failure");
+        }
     }
 
     private void copyDtoToEntity(Client entity, ClientDTO dto) {
